vec3 module setup

+++ src/math.rs
@@ -10,15 +10,15 @@ use std::ops::{Mul,Div};
 // allows println to show struct - https://doc.rust-lang.org/book/ch05-02-example-structs.html
 #[derive(Debug)]
 pub struct Vec3{
-	pub x: f32,
-	pub y: f32,
-	pub z: f32,
+	pub x: f64,
+	pub y: f64,
+	pub z: f64,
 }
 
-impl Mul<f32> for Vec3 {
+impl Mul<f64> for Vec3 {
 	type Output = Vec3;
 
-	fn mul(self, other: f32) -> Vec3 {
+	fn mul(self, other: f64) -> Vec3 {
 		Vec3 { 
 			x: self.x * other,
 			y: self.y * other,
@@ -27,10 +27,10 @@ impl Mul<f32> for Vec3 {
 	}
 }
 
-impl Div<f32> for Vec3 {
+impl Div<f64> for Vec3 {
 	type Output = Vec3;
 
-	fn div(self, other: f32) -> Vec3 {
+	fn div(self, other: f64) -> Vec3 {
 		Vec3 { 
 			x: self.x / other,
 			y: self.y / other,
@@ -41,7 +41,7 @@ impl Div<f32> for Vec3 {
 
 
 impl Vec3 {
-	pub fn new(sx: f32, sy: f32, sz: f32) -> Vec3 {
+	pub fn new(sx: f64, sy: f64, sz: f64) -> Vec3 {
 		Vec3 {x: sx, y: sy, z: sz}
 	}
 
@@ -54,7 +54,7 @@ impl Vec3 {
 	}
 	
 	// borrows the vector to return another value
-	pub fn length(&self) -> f32 {
+	pub fn length(&self) -> f64 {
 		let val = self.x*self.x + self.y*self.y + self.z*self.z;
 		val.sqrt()
 	}
